Create the services and improve the performance about chrome memory.
